test(topology): add test for TopologyUtil
